backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
spacy>=3.7.0
transformers>=4.44.0
gensim>=4.3.0
opencv-python-headless>=4.10.0
pytrends>=4.9.2
google-genai>=1.5.0
numpy>=1.26.0,<2.0
pandas>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
sentence-transformers>=3.0.0
textstat>=0.7.0
scikit-learn>=1.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
pytest-timeout>=2.3.0
